refactored some static methods to extension methods.
